Use remote builders to provide native arm64 builds
This temporarily uses redpencil/plugin-docker-buildx instead of the
upstream image from woodpeckerci. To be changed back if and when the
plugin gets merged upstream.
The woodpecker server will need secrets setup for "ssh_key" and
"remote_builders", and the key added to the builder's authorized_keys
file. Not sure this was done globally already or per org.